# docs/api-docs.yaml
openapi: 3.0.3
info:
  title: Registry Viewer API
  description: API for registry viewer to obtain registry data
  contact:
    email: myung.choi@gtri.gatech.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: v1.11.1
servers:
  - url: https://smartchartsuite.dev.heat.icl.gtri.org
    description: Production server
  - url: http://localhost:8080
    description: Local development (optional)
...
paths:
...
  /registry-viewer-api/metadata:
    get:
      tags:
        - metadata-api-controller
      summary: Getting Registry Information
      description: Obtaining information about the availabe registries.
      operationId: getMetadataUsingGET
      responses:
        '200':
          description: List of registry metadata entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadatas'
              example:
                count: 1
                metadatas:
                  - name: Syphilis Registry
                    description: Registry description
                    tag: syphilis
                    viewerConfig: {}
                version: 1.0.0
        '400':
          description: Bad request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - Authorization: []
        - basicAuth: []
    put:
      tags:
        - metadata-api-controller
      summary: Editing Registry Information
      description: Editing registry information.
      operationId: editMetadataUsingPUT

      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
      responses:
        '200':
          description: The registry information edited
        '400':
          description: Bad request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - Authorization: []
        - basicAuth: []
    post:
      tags:
        - metadata-api-controller
      summary: Adding Registry Information
      description: Adding new registry information.
      operationId: addMetadataUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
      responses:

        '400':
          description: Bad request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '409':
          description: Conflict – entry with same tag already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
      security:
        - Authorization: []
        - basicAuth: []


    delete:
      tags:
        - metadata-api-controller
      summary: Delete Registry Information
      description: Delete a metadata entry by its tag.
      operationId: deleteMetadataUsingDELETE
      parameters:
        - name: tag
          in: query
          description: Tag of the metadata to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Metadata entry successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '404':
          description: Metadata entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '400':
          description: Bad request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - Authorization: []
        - basicAuth: []
...
components:
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header
      description: API key passed in the `Authorization` header (e.g., `Bearer <token>`)
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication (optional)
  schemas:
    Error:
      type: object
      description: Generic error response
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: Invalid request parameters
      required:
        - code
        - message
...
    JsonNode:
      title: JsonNode
      type: object
...
    Metadata:
      title: Metadata
      type: object
      properties:
        description:
          type: string
          example: Registry description
        name:
          type: string
          example: Syphilis Registry
        tag:
          type: string
          example: Registry tag name
        viewerConfig:
          $ref: '#/components/schemas/JsonNode'
          example: Registry Viewer UI Config
    Metadatas:
      title: Metadatas
      type: object
      properties:
        count:
          type: integer
          format: int32
          example: 1
        metadatas:
          type: array
          items:
            $ref: '#/components/schemas/Metadata'
        version:
          type: string
          example: 1.0.0
...